Display navbar differently
Navbar won't be a simple element with a fixed position because it forced
me to adding padding to every content on the page (even if the navbar
wasn't included). Therefore I chose to put both, navbar and content in
flexbox container (root of app) so they don't depend
on each other as much. Now the content can have padding of any size
and it's independent of width of the side navbar.